Implement suggestion from pull request discussion.
